+++ govtool/frontend/src/consts/CIP100Context.ts
@@ -1,0 +1,51 @@
+export const CIP_100_CONTEXT = {
+  "@language": "en-us",
+  CIP100:
+    "https://github.com/cardano-foundation/CIPs/blob/master/CIP-0100/README.md#",
+  hashAlgorithm: "CIP100:hashAlgorithm",
+  body: {
+    "@id": "CIP100:body",
+    "@context": {
+      references: {
+        "@id": "CIP100:references",
+        "@container": "@set" as const,
+        "@context": {
+          GovernanceMetadata: "CIP100:GovernanceMetadataReference",
+          Other: "CIP100:OtherReference",
+          label: "CIP100:reference-label",
+          uri: "CIP100:reference-uri",
+          referenceHash: {
+            "@id": "CIP100:referenceHash",
+            "@context": {
+              hashDigest: "CIP100:hashDigest",
+              hashAlgorithm: "CIP100:hashAlgorithm",
+            },
+          },
+        },
+      },
+      comment: "CIP100:comment",
+      externalUpdates: {
+        "@id": "CIP100:externalUpdates",
+        "@context": {
+          title: "CIP100:update-title",
+          uri: "CIP100:uri",
+        },
+      },
+    },
+  },
+  authors: {
+    "@id": "CIP100:authors",
+    "@container": "@set" as const,
+    "@context": {
+      name: "http://xmlns.com/foaf/0.1/name",
+      witness: {
+        "@id": "CIP100:witness",
+        "@context": {
+          witnessAlgorithm: "CIP100:witnessAlgorithm",
+          publicKey: "CIP100:publicKey",
+          signature: "CIP100:signature",
+        },
+      },
+    },
+  },
+};